trying to simplify my SRBlip object
